# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server CRM server.",
        "title": "Golang CRM Swagger Documentation",
        "contact": {
            "name": "API Support",
            "url": "https://t.me/xalmatoff",
            "email": "venom.uzz@mail.ru"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/product-images": {
            "post": {
                "description": "This API to create a productImage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductImages"
                ],
                "summary": "Create a productImage.",
                "parameters": [
                    {
                        "type": "string",
                        "x-order": "1",
                        "name": "productId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "x-order": "2",
                        "name": "position",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "fileImage",
                        "name": "fileImage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ProductImages"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/product-images/{id}": {
            "get": {
                "description": "This API to get productImage by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductImages"
                ],
                "summary": "Get productImage by id.",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "id",
                        "description": "id for get ProductImage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductImages"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "This API to update a productImage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductImages"
                ],
                "summary": "Update a productImage.",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "id",
                        "description": "id for update ProductImage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-order": "1",
                        "name": "productId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "x-order": "2",
                        "name": "position",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "file_image",
                        "name": "file_image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductImages"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "This API to delete a productImage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductImages"
                ],
                "summary": "Delete a productImage.",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "id",
                        "description": "id for delete ProductImage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ProductImages": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "1"
                },
                "productId": {
                    "type": "string",
                    "x-order": "2"
                },
                "image": {
                    "type": "string",
                    "x-order": "3"
                },
                "position": {
                    "type": "integer",
                    "x-order": "4"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "5"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "6"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "7"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}